# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for MOOVON Consulting
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lead {
  id: number;
  name: string;
  email: string;
  whatsapp: string;
  createdAt: string;
}

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  whatsapp: string;
}

export interface Contact {
  id: number;
  name: string;
  phone: string;
  email: string;
  message: string;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  phone: string;
  email: string;
  /** @minLength 10 */
  message: string;
}

export interface ApiError {
  error: string;
}