# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for MOOVON Consulting
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  phone: string;
  email: string;
  /** @minLength 10 */
  message: string;
}